refactor: merge twopoint_convert + one_covariance_io into sacc_interop
One module whose job is converting between the SACC product and
external analysis-tool file formats (CosmoSIS 2pt-FITS, OneCovariance).
Public API unchanged; sacc_io untouched.
